+++ test/functional/thread_safety_test.rb
@@ -6,7 +6,7 @@ class ThreadSafetyTest < ActionController::TestCase
 
     slow_thread = Thread.new do
       controller = TestController.new
-      controller.send :set_whodunnit
+      controller.send :set_paper_trail_whodunnit
       begin
         sleep 0.001
       end while blocked
@@ -15,7 +15,7 @@ class ThreadSafetyTest < ActionController::TestCase
 
     fast_thread = Thread.new do
       controller = TestController.new
-      controller.send :set_whodunnit
+      controller.send :set_paper_trail_whodunnit
       who = PaperTrail.whodunnit
       blocked = false
       who